+++ egov/egov-eis/src/main/java/org/egov/pims/dao/PersonalInformationDAO.java
@@ -67,7 +67,7 @@ public interface PersonalInformationDAO extends org.egov.infstr.dao.GenericDAO
 	public void deleteLangKnownForEmp(PersonalInformation personalInformation);
 	public List getListOfPersonalInformationByEmpIdsList(List empIdsList);
 	public PersonalInformation getEmployee(Integer deptId, Integer designationId, Long Boundaryid)throws TooManyValuesException, NoSuchObjectException;
-	public PersonalInformation getEmployeeByFunctionary(Integer deptId, Integer designationId, Long Boundaryid,Integer functionaryId)throws TooManyValuesException, NoSuchObjectException;
+	public PersonalInformation getEmployeeByFunctionary(Long deptId, Long designationId, Long Boundaryid,Integer functionaryId)throws TooManyValuesException, NoSuchObjectException;
 	public List getListOfUsersByBoundaryId(Long boundaryId) throws NoSuchObjectException;
 	public List getListOfUsersForGivenBoundaryId(Long boundaryId) throws NoSuchObjectException;
 	/**
